requirements.txt
# Sentinel Graph Initialization & Ingestion dependencies
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
neo4j>=5.20.0,<6.0.0
langchain-core>=0.2.0,<0.4.0
langchain-groq>=0.1.9,<0.3.0
langchain-huggingface>=0.1.0,<1.0.0
sentence-transformers>=3.0.1,<4.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
streamlit>=1.36.0,<2.0.0
google-genai>=0.8.0,<2.0.0
llama-cpp-python>=0.2.90,<1.0.0